% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorizeLAS}
\alias{colorizeLAS}
\title{Colorize LAS Point Cloud Based on Attribute Values}
\usage{
colorizeLAS(las, attribute, palette_func, n = 100)
}
\arguments{
\item{las}{LAS object containing the point cloud data}

\item{attribute}{character Name of the attribute column in the LAS data to use for coloring}

\item{palette_func}{function Color palette function that generates colors (e.g., rainbow, heat.colors, lidR::random.colors)}

\item{n}{integer Number of colors to generate in the palette (default: 100)}
}
\value{
LAS object with RGB color channels (R, G, B) added based on the specified attribute
}
\description{
This function colorizes a LAS point cloud by mapping attribute values to RGB colors
using a specified color palette function. The function normalizes the attribute values and applies
the color mapping to create RGB color channels in the LAS object.
}
\details{
The function handles NA values by replacing them with the minimum attribute value.
Attribute values are normalized to the range 0-1 before mapping to colors. The resulting
RGB values are added as new columns (R, G, B) to the LAS object.
}
\examples{
\dontrun{
# Load LAS data
las <- readLAS("path/to/file.las")

# Color by height using rainbow palette
las_colored <- colorizeLAS(las, "Z", rainbow, n = 256)

# Color by intensity using heat colors
las_colored <- colorizeLAS(las, "Intensity", heat.colors, n = 100)

# Color by tree ID using random colors
las_colored <- colorizeLAS(las, "treeID", lidR::random.colors, n = 50)
}

}
