% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFits.R
\name{fit_circle}
\alias{fit_circle}
\title{Fit a circle to a point cloud}
\usage{
fit_circle(pts_mat, method = "ransac", n = 20, p = 0.9, P = 0.99, c0 = T)
}
\arguments{
\item{pts_mat}{A matrix of the point cloud coordinates.}

\item{method}{The method to use for fitting. Either 'ransac' or 'leastsq'.}

\item{n}{The number of points sampled in every RANSAC iteration. Only used if method is 'ransac'.}

\item{p}{The proportion of inliers in P. Only used if method is 'ransac'.}

\item{P}{The level of confidence desired. Only used if method is 'ransac'.}

\item{c0}{If TRUE, centers x and y to zero. Only used if method is 'leastsq'.}
}
\value{
A vector with four elements: The x and y coordinates of the center of the circle, the estimated radius of the circle, and the root mean square error (rmse) of the distances from the points to the circle.
}
\description{
This function fits a circle to a point cloud using either the RANSAC or least squares method.
The function returns the estimated parameters of the circle, the root mean square error (RMSE) of the distances from the points to the circle, and the x, y coordinates of the center of the circle.
}
\examples{
 \dontrun{
# Create a set of points on a circle with center (2, 3) and radius 5
theta <- seq(0, 2*pi, length.out = 100)
xp <- 2 + 5 * cos(theta)
yp <- 3 + 5 * sin(theta)
pts_mat <- cbind(xp, yp)

# Use fit_circle to estimate the center, radius and RMSE
result <- fit_circle(pts_mat, method = 'ransac', n=15, p=.8, P=.99)
}
}
