% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinderFits.R
\name{fit_cylinder}
\alias{fit_cylinder}
\title{Fit a cylinder to a point cloud}
\usage{
fit_cylinder(
  pts_mat,
  method = "ransac",
  select_n = 20,
  p = 0.9,
  P = 0.99,
  timesN = 5,
  init = NULL,
  opt.method = "Nelder-Mead",
  max.iter = 10,
  rmse.threshold = 0.001,
  per.allowable.error = 50,
  cores = detectCores(logical = F),
  m.estimator = "tukey"
)
}
\arguments{
\item{pts_mat}{A matrix of the point cloud coordinates.}

\item{method}{The method to use for fitting. Either 'ransac' or 'leastsq'.}

\item{select_n}{The number of points sampled in every RANSAC iteration.}

\item{p}{The proportion of inliers in P. Only used if method is 'ransac'.}

\item{P}{The level of confidence desired. Only used if method is 'ransac'.}

\item{timesN}{An inflation factor to multiply the number of RANSAC iterations by. Only used if method is 'ransac'.}

\item{init}{An initial guess of cylinder parameters.}

\item{opt.method}{The optimization method to use. Passed on to optim.}

\item{max.iter}{The maximum number of iterations for the optimization algorithm. Only used if method is 'leastsq'.}

\item{rmse.threshold}{The threshold for the root mean square error (RMSE) for the optimization algorithm to stop.}

\item{per.allowable.error}{The percentage of allowable error for the RMSE.}

\item{cores}{The number of cores to use for parallel computing. Only used if method is 'ransac'.}

\item{m.estimator}{The type of M-estimator to use for robust estimation. Only used if method is 'leastsq'.}
}
\value{
A vector with nine elements: The estimated parameters of the cylinder (rho, theta, phi, alpha, radius), the root mean square error (rmse) of the distances from the points to the cylinder, and the x, y, and z coordinates of the center of the base of the cylinder.
}
\description{
This function fits a cylinder to a point cloud using either the RANSAC or least squares method.
The function returns the estimated parameters of the cylinder, the root mean square error (RMSE) of the distances from the points to the cylinder, and the x, y, and z coordinates of the center of the base of the cylinder.
}
\examples{
 \dontrun{
# Define the cylinder attributes
npts = 500
cyl_length = 0.5
radius = 0.2718

# Generate the X,Y,Z values
Z=runif(n = n, min = 0, max = cyl_length)
angs = runif(n, 0, 2*pi)
X = sin(angs)*radius
Y = cos(angs)*radius

# Creation of a LAS object out of external data
cloud <- LAS(data.frame(X,Y,Z))

# Fit a cylinder and retrun the information
cyl_par = fit_cylinder(cloud, method = 'ransac', n=5, p=.9, P=.95, timesN=5, opt.method='Nelder-Mead', select_n=20, max.iter=10, rmse.threshold=0.001, per.allowable.error=50, cores=1, m.estimator='tukey')
}
}
