% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{A simple function to download some example MLS data in a ponderosa pine forest north of Flagstaff AZ}
\usage{
getExampleData()
}
\description{
A simple function to download some example MLS data in a ponderosa pine forest north of Flagstaff AZ
}
