% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFits.R
\name{least.sq.circle}
\alias{least.sq.circle}
\title{Fit a circle to a point cloud}
\usage{
least.sq.circle(xp, yp, c0 = T)
}
\arguments{
\item{xp}{A vector of the x coordinates of the point cloud.}

\item{yp}{A vector of the y coordinates of the point cloud.}

\item{c0}{If TRUE, centers x and y to zero.}
}
\value{
A vector with four elements: The x and y coordinates of the center of the circle, the estimated radius of the circle, and the root mean square error (rmse) of the distances from the points to the circle.
}
\description{
This function fits a circle to a point cloud using the least squares method.
The function returns the estimated parameters of the circle, the root mean square error (RMSE) of the distances from the points to the circle, and the x and y coordinates of the center of the circle.
}
\examples{
 \dontrun{
# Create a set of points on a circle with center (2, 3) and radius 5
theta <- seq(0, 2*pi, length.out = 100)
xp <- 2 + 5 * cos(theta)
yp <- 3 + 5 * sin(theta)

# Use circlefit to estimate the center, radius and RMSE
result <- circlefit(xp, yp)
}
}
