% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_Morph.R
\name{plot_raster_by_name_base}
\alias{plot_raster_by_name_base}
\title{Fallback function for base plotting when ggplot2 is not available}
\usage{
plot_raster_by_name_base(rasters, name)
}
\description{
Fallback function for base plotting when ggplot2 is not available
}
\keyword{internal}
