% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_tree_detection.R
\name{plot_tree_detection}
\alias{plot_tree_detection}
\title{Plot Tree Detection Results}
\usage{
plot_tree_detection(las, trees_sf, color_by = "Quality", point_size = 3)
}
\arguments{
\item{las}{Original LAS object}

\item{trees_sf}{sf object from comprehensive_tree_detection}

\item{color_by}{Color trees by this attribute}

\item{point_size}{Size of tree markers}
}
\value{
ggplot object
}
\description{
Creates a visualization of detected trees overlaid on the point cloud
}
