% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_Morph.R
\name{process_rasters_patchmorph}
\alias{process_rasters_patchmorph}
\title{Process rasters based on suitability, gap, and spur parameters}
\usage{
process_rasters_patchmorph(input_raster, suitList, gapList, spurList)
}
\arguments{
\item{input_raster}{RasterLayer The input raster to be processed.}

\item{suitList}{numeric A vector of suitability levels for reclassification.}

\item{gapList}{numeric A vector of gap distances for processing.}

\item{spurList}{numeric A vector of spur distances for processing.}
}
\value{
list A list of processed rasters with names indicating the suitability, gap, and spur parameters used.
}
\description{
\code{process_rasters_patchmorph} processes an input raster by reclassifying it based on suitability levels,
and then applying gap and spur distance transformations to generate a list of processed rasters.
}
\examples{
\dontrun{
# Define input parameters
input_raster <- raster::raster(system.file("external/test.grd", package="raster"))
suitList <- c(0.2, 0.5, 0.8)
gapList <- c(1, 2, 3)
spurList <- c(0.5, 1, 1.5)

# Process the rasters
processed_rasters <- process_rasters_patchmorph(input_raster, suitList, gapList, spurList)

# Plot the first processed raster
plot(processed_rasters[[1]])
}

}
