% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_Morph.R
\name{process_single_combination}
\alias{process_single_combination}
\title{Process single parameter combination for patch morphology}
\usage{
process_single_combination(
  input_raster,
  suit,
  gapDist,
  spurDist,
  classify_matrices,
  get_kernel_func,
  raster_res
)
}
\description{
Process single parameter combination for patch morphology
}
\keyword{internal}
