% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFits.R
\name{ransac.circle}
\alias{ransac.circle}
\title{RAndom SAmple Consensus (RANSAC) circle fit}
\usage{
ransac.circle(slice, n = 15, p = 0.8, P = 0.99)
}
\arguments{
\item{slice}{A matrix of the point cloud coordinates.}

\item{n}{The number of points to sample on every RANSAC iteration.}

\item{p}{The estimated proportion of inliers in the dataset.}

\item{P}{The level of confidence desired.}
}
\value{
A vector with four elements: The x and y coordinates of the center of the circle, the estimated radius of the circle, and the root mean square error (rmse) of the distances from the points to the circle.
}
\description{
This function fits a circle to a point cloud using the RANSAC algorithm.
The function returns the estimated parameters of the circle, the root mean square error (RMSE) of the distances from the points to the circle, and the x and y coordinates of the center of the circle.
}
\examples{
 \dontrun{
# Create a set of points on a circle with center (2, 3) and radius 5
theta <- seq(0, 2*pi, length.out = 100)
xp <- 2 + 5 * cos(theta)
yp <- 3 + 5 * sin(theta)
slice <- cbind(xp, yp)

# Use RANSAC.circle to estimate the center, radius and RMSE
result <- RANSAC.circle(slice, n=15, p=.8, P=.99)
}
}
