% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinderFits.R
\name{ransac.cylinder}
\alias{ransac.cylinder}
\title{Fit a cylinder to a point cloud using RANSAC}
\usage{
ransac.cylinder(
  stem.sec,
  select_n = 20,
  p = 0.9,
  P = 0.99,
  timesN = 5,
  init = NULL,
  opt.method = "Nelder-Mead",
  rmse.threshold = 0.001,
  cores = detectCores(logical = F)
)
}
\arguments{
\item{stem.sec}{A matrix of the point cloud coordinates.}

\item{select_n}{The number of points sampled in every RANSAC iteration.}

\item{p}{The proportion of inliers in stem.sec.}

\item{P}{The level of confidence desired.}

\item{timesN}{An inflation factor to multiply the number of RANSAC iterations by.}

\item{init}{An initial guess of cylinder parameters.}

\item{opt.method}{The optimization method to use. Passed on to optim.}

\item{rmse.threshold}{The threshold for the root mean square error (RMSE) for the optimization algorithm to stop.}

\item{cores}{The number of cores to use for parallel computing. Defaults to the number of cores available on the machine.}
}
\value{
A named vector with nine elements: The estimated parameters of the cylinder (rho, theta, phi, alpha, radius), the root mean square error (rmse) of the distances from the points to the cylinder, and the x, y, and z coordinates of the center of the base of the cylinder.
}
\description{
This function fits a cylinder to a point cloud using the RANSAC method. It uses parallel computing to speed up the RANSAC iterations.
}
\examples{
 \dontrun{
# Define the cylinder attributes
npts = 500
cyl_length = 0.5
radius = 0.2718

# Generate the X,Y,Z values
Z=runif(n = n, min = 0, max = cyl_length)
angs = runif(n, 0, 2*pi)
X = sin(angs)*radius
Y = cos(angs)*radius

# Creation of a LAS object out of external data
cloud <- LAS(data.frame(X,Y,Z))

# Fit a cylinder and retrun the information
cyl_par = ransac.cylinder(cloud, n=5, p=.9, P=.95, timesN=5, opt.method='Nelder-Mead', rmse.threshold=0.001, cores=1)
}
}
