\name{read.excel}
\alias{read.excel}
\title{Copies data from clipboard, e.g. when pasting from Excel.}
\usage{
  read.excel(header = TRUE, ...)
}
\arguments{
  \item{header}{If TRUE, reads data with header row}

  \item{...}{Passed to \code{\link{read.table}}}
}
\description{
  Copies data from clipboard, e.g. when pasting from Excel.
}
\seealso{
  \code{\link{read.excel}} and the base functions
  \code{\link{read.table}}, \code{\link{readClipboard}}

  Other Functions.that.copy.paste:
  \code{\link{write.excel}}, \code{\link{write.ppt}}
}

