\name{spanner-package}
\alias{spanner-package}
\alias{spanner}
\docType{package}
\title{
\packageTitle{spanner}
}
\description{
\packageDescription{spanner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spanner}
\packageIndices{spanner}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spanner}

Maintainer: \packageMaintainer{spanner}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<lidR>:<lidR>-package]{<lidR>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
