% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_features_optimized.R
\name{test_geometric_features}
\alias{test_geometric_features}
\title{Test the simple geometric features function}
\usage{
test_geometric_features(las, radius, max_neighbors = 50)
}
\arguments{
\item{las}{LAS object from lidR package}

\item{radius}{Search radius for neighbor computation}

\item{max_neighbors}{Maximum number of neighbors to consider}
}
\value{
List of geometric features
}
\description{
Test the simple geometric features function
}
