% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raster_Eigen_TreeLocations.R
\name{get_raster_eigen_treelocs}
\alias{get_raster_eigen_treelocs}
\title{Obtain tree information by rasterizing point cloud values of relative neighborhood density and 
verticality within a slice of a normalized point cloud}
\usage{
get_raster_eigen_treelocs(
  las = las,
  res = 0.05,
  pt_spacing = 0.0254,
  dens_threshold = 0.2,
  eigen_threshold = 0.6666,
  grid_slice_min = 0.6666,
  grid_slice_max = 2,
  grid_slice_buffer = 0.5,
  minimum_polygon_area = 0.025,
  cylinder_fit_type = "ransac",
  output_location = getwd()
)
}
\arguments{
\item{las}{LAS Description to go here.}

\item{res}{numeric Description to go here.}

\item{pt_spacing}{numeric Description to go here.}

\item{dens_threshold}{numeric Description to go here.}

\item{eigen_threshold}{numeric Description to go here.}

\item{grid_slice_min}{numeric Description to go here.}

\item{grid_slice_max}{numeric Description to go here.}

\item{grid_slice_buffer}{numeric Description to go here.}

\item{minimum_polygon_area}{numeric Description to go here.}

\item{cylinder_fit_type}{character "ransac" or "irls"}

\item{output_location}{character Description to go here.}

\item{plot_radius}{numeric Description to go here.}
}
\value{
data.frame A data.frame contained the following seed information: `TreeID`,
`X`, `Y`, `Z`, and `Radius` in the same units as the .las
}
\description{
`get_raster_eigen_treelocs` returns a data.frame containing `TreeID`, `X`, `Y`, `Z`, `Radius` 
and `Error` in the same units as the .las
}
\details{
For terrestrial and mobile lidar datasets, tree locations and estimates of DBH are provided
by rasterizing individual point cloud values of relative neighborhood density (at 0.3 and 
1 m radius) and verticality within a slice of the normalized point cloud around breast height
(1.34 m). The algorithim then uses defined threshold values to classify the resulting rasters
and create unique polygons from the resulting classified raster. These point-density and 
verticality polygons were selected by their intersection with one another, resulting in a
final set of polygons which were used to clip out regions of the point cloud that were most 
likely to represent tree boles. A RANSAC cylinder fitting algorithm was then used to estimate
the fit of a cylinder to individual bole points. Cylinder centers and radius were used as inputs
to an individual tree segmentation
}
\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F, F, F, T, T)

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{
sum("a")
}

}
