% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{A simple function to download some example MLS/TLS data in a ponderosa pine forest north of Flagstaff AZ}
\usage{
getExampleData(
  lazName = c("PineExampleA", "DensePatchA", "DensePatchB", "SparsePatchA",
    "SparsePatchB", "ZebcamExample")
)
}
\arguments{
\item{character}{The basename of the example dataset to be downloaded. Possibilities include:
\itemize{
\item "PineExampleA" - A 1ha exmaple dataset pre-downsampled using a 0.01 cm voxel grid
\item "DensePatchA" - A small, dense ponderosa pine patch with many, small trees
\item "DensePatchB"  - A small, dense ponderosa pine patch with few, medium-sized trees
\item "SparsePatchA"  - A small, open ponderosa pine patch with few, medium-sized trees
\item "SparsePatchB" - A small, open ponderosa pine patch with few, largely spaced trees
\item "ZebcamExample" - A big, RGB colored dataset
}}
}
\value{
Nothing Downloads and saves a specified .laz file to the "extdata" folder of the 'spanner' package
}
\description{
Note: none of these datasets have been downsampled, classified for ground points, or normalized, all of which
must be done before using them with 'Raster_Eigen_TreeLocations' or 'Segment_Graph' functions in the 'spanner' package.
}
\examples{


\dontrun{
# Specify one of six possible datasets to download
getExampleData("SparsePatchA")
}
}
