% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_Morph.R
\name{plot_raster_by_name}
\alias{plot_raster_by_name}
\title{Plot a raster by its name}
\usage{
plot_raster_by_name(rasters, raster_name)
}
\arguments{
\item{rasters}{list A list of rasters.}

\item{raster_name}{character The name of the raster to be plotted.}
}
\value{
NULL This function does not return a value. It plots the raster if found.
}
\description{
\code{plot_raster_by_name} plots a raster from a list of rasters based on the provided raster name.
}
\examples{
\dontrun{
# Define a list of rasters
rasters <- list(raster1 = raster::raster(matrix(1:100, 10, 10)),
                raster2 = raster::raster(matrix(101:200, 10, 10)))

# Assign names to the rasters
names(rasters[[1]]) <- "raster1"
names(rasters[[2]]) <- "raster2"

# Plot a raster by its name
plot_raster_by_name(rasters, "raster1")
}

}
