% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch_Morph.R
\name{sum_rasters_by_suitability}
\alias{sum_rasters_by_suitability}
\title{Sum rasters by suitability level}
\usage{
sum_rasters_by_suitability(rasters, suitList)
}
\arguments{
\item{rasters}{list A list of rasters.}

\item{suitList}{numeric A vector of suitability levels.}
}
\value{
list A list of summed rasters for each suitability level.
}
\description{
\code{sum_rasters_by_suitability} sums rasters from a list based on their suitability levels.
}
\examples{
\dontrun{
# Define a list of rasters
rasters <- list(raster1 = raster::raster(matrix(1:100, 10, 10)),
                raster2 = raster::raster(matrix(101:200, 10, 10)))

# Assign names to the rasters
names(rasters[[1]]) <- "suit_0.2"
names(rasters[[2]]) <- "suit_0.5"

# Define suitability levels
suitList <- c(0.2, 0.5)

# Sum rasters by suitability level
summed_rasters <- sum_rasters_by_suitability(rasters, suitList)
}

}
